refactor devices (#1699)
* refactor devices

* add devices

* remove bad import
