+++ composer/algorithms/seq_length_warmup/seq_length_warmup.py
@@ -15,9 +15,9 @@ from composer.core import Algorithm, Event, State
 from composer.core.precision import get_precision_context
 from composer.core.time import TimeUnit
 from composer.core.types import Batch
+from composer.devices import DeviceGPU
 from composer.loggers import Logger
 from composer.models import HuggingFaceModel
-from composer.trainer.devices import DeviceGPU
 from composer.utils import dist, ensure_tuple
 
 log = logging.getLogger(__name__)
